+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`0.14.2...master`][0.14.2...master].
+For a full diff see [`0.14.3...master`][0.14.3...master].
+
+## [`0.14.3`][0.14.3]
+
+For a full diff see [`0.14.2...0.14.3`][0.14.2...0.14.3].
+
+### Fixed
+
+* Ignored first line in `DeclareStrictTypesRule` when it is a shebang ([#186]), by [@Great-Antique]
 
 ## [`0.14.2`][0.14.2]
 
@@ -296,6 +304,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [0.14.0]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.14.0
 [0.14.1]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.14.1
 [0.14.2]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.14.2
+[0.14.3]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.14.3
 
 [362c7ea...0.1.0]: https://github.com/ergebnis/phpstan-rules/compare/362c7ea...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/phpstan-rules/compare/0.1.0...0.2.0
@@ -318,7 +327,8 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [0.13.0...0.14.0]: https://github.com/ergebnis/phpstan-rules/compare/0.13.0...0.14.0
 [0.14.0...0.14.1]: https://github.com/ergebnis/phpstan-rules/compare/0.14.0...0.14.1
 [0.14.1...0.14.2]: https://github.com/ergebnis/phpstan-rules/compare/0.14.1...0.14.2
-[0.14.2...master]: https://github.com/ergebnis/phpstan-rules/compare/0.14.2...master
+[0.14.2...0.14.3]: https://github.com/ergebnis/phpstan-rules/compare/0.14.2...0.14.3
+[0.14.3...master]: https://github.com/ergebnis/phpstan-rules/compare/0.14.3...master
 
 [#1]: https://github.com/ergebnis/phpstan-rules/pull/1
 [#4]: https://github.com/ergebnis/phpstan-rules/pull/4
@@ -365,6 +375,8 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#158]: https://github.com/ergebnis/phpstan-rules/pull/158
 [#161]: https://github.com/ergebnis/phpstan-rules/pull/161
 [#166]: https://github.com/ergebnis/phpstan-rules/pull/166
+[#186]: https://github.com/ergebnis/phpstan-rules/pull/186
 
 [@ergebnis]: https://github.com/ergebnis
+[@Great-Antique]: https://github.com/Great-Antique
 [@localheinz]: https://github.com/localheinz